List below: 1. fixed database to work with abstract class Employee 2. added URL getters to the db 2. add URLs for creating records in the database

+++ HotelManagement/src/main/java/com/hotelmangementprogram/hotelmanagement/model/Room.java
@@ -17,7 +17,7 @@ public class Room {
     @Column(name = "roomId")
     private Long roomId;
     @Column(name = "roomNumber")
-    private int roomNumber;
+    private Long roomNumber;
     @Column(name = "roomPrice")
     private double roomPrice;
     @Column(name = "roomIsEmpty")
